% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{apply_to_dataframe}
\alias{apply_to_dataframe}
\title{Apply a function to every variable of a data frame.}
\usage{
apply_to_dataframe(data, column_function)
}
\arguments{
\item{data}{Data Frame to be analyzed}

\item{column_function}{Function to be applied to each varaible.}
}
\value{
Vector of results.
}
\description{
Apply a function to every variable of a data frame.
}
