% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{average_missing}
\alias{average_missing}
\title{Dataframe function - measure the average rate of missingness on a column}
\usage{
average_missing(data, missing_types = MISSING_TYPES)
}
\arguments{
\item{data}{The dataframe to measure missingness from}

\item{missing_type}{Vector of values classified as missing.}
}
\value{
Vector of prevalence of missing values.
}
\description{
Dataframe function - measure the average rate of missingness on a column
}
