% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{average_missing_heuristic}
\alias{average_missing_heuristic}
\title{Check if missing data rate is less than set limit.}
\usage{
average_missing_heuristic(data, limit = 0.1, missing_types = MISSING_TYPES)
}
\arguments{
\item{data}{Data.frame of variables to be analyzed}

\item{limit}{Allowable rate of missingness, valued [0, 1]}

\item{missing_types}{[optional] vector of valuse to define as missing.}
}
\value{
Boolean-vector - FALSE: missing more data than the limit.
}
\description{
Check if missing data rate is less than set limit.
}
