% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{blank_result}
\alias{blank_result}
\title{Generate a blank result vector from a data frame}
\usage{
blank_result(data)
}
\arguments{
\item{data}{The data frame to template the result vector from.}
}
\description{
Generate a blank result vector from a data frame
}
