% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_mining.R
\name{build_model}
\alias{build_model}
\title{Constuct model via `caret` call}
\usage{
build_model(data, method = "rpart")
}
\arguments{
\item{data}{Data frame for analysis}

\item{method}{Method call from `carat::train` to use for binary-classication}
}
\value{
`caret` model
}
\description{
Constuct model via `caret` call
}
