% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutual_info.R
\name{column_info_func}
\alias{column_info_func}
\title{Function factory for mutual information.}
\usage{
column_info_func(col1, col2)
}
\arguments{
\item{col1}{Name of the first column}

\item{col2}{Name of the second column}
}
\description{
Creates a function to be called in the tidyverse that calculates 
mutual information between variables of a data set.  Intended to be used as 
part of a group_by(...) %>% group_map(column_info_func(...)) workflow.
}
\examples{
ChickWeight \%>\% group_by(Diet)  \%>\% 
    group_map(column_info_func("Chick", "weight")) \%>\% 
    do.call(rbind, .)

}
