% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R, R/miscoding_prescriptive.R
\name{compare}
\alias{compare}
\title{Apply a comparison function to a column of data}
\usage{
compare(data, col_name, validation)

compare(data, col_name, validation)
}
\arguments{
\item{data}{The data set to be analyzed}

\item{col_name}{The column of data to be considered}

\item{validation}{A list of validation types.}
}
\value{
quantity of data that obeys validation rule,  valued [0, 1].
}
\description{
Apply a comparison function to a column of data

Apply a comparison function to a column of data
}
