% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare2.R, R/miscoding_explorative.R
\name{compare_classes}
\alias{compare_classes}
\title{Checks if there are enough examples for the number of classes in a variable.}
\usage{
compare_classes(values, comparison = sqrt_comparison)

compare_classes(values, comparison = sqrt_comparison)
}
\arguments{
\item{values}{Vector of observations}

\item{comparison}{[optional] A function to compare the number of unique classes with the number of observations.}
}
\value{
Boolean - FALSE: not enough examples present

Boolean - FALSE: not enough examples present
}
\description{
Checks if there are enough examples for the number of classes in a variable.

Checks if there are enough examples for the number of classes in a variable.
}
