% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare2.R, R/miscoding_explorative.R
\name{compare_frame_classes}
\alias{compare_frame_classes}
\title{Check each variable of a data frame to determine if there are enough examples for the number of distinct values.  
NB: will return a spurious result for continous valued variables.}
\usage{
compare_frame_classes(data, comparison = sqrt_comparison)

compare_frame_classes(data, comparison = sqrt_comparison)
}
\arguments{
\item{data}{Dataframe of variables for checking}

\item{comparison}{[optional] A function to compare the number of unique classes with the number of observations.}
}
\value{
Boolean-vector -  FALSE: not enough examples present

Boolean-vector -  FALSE: not enough examples present
}
\description{
Check each variable of a data frame to determine if there are enough examples for the number of distinct values.  
NB: will return a spurious result for continous valued variables.

Check each variable of a data frame to determine if there are enough examples for the number of distinct values.  
NB: will return a spurious result for continous valued variables.
}
