% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare2.R, R/miscoding_explorative.R
\name{compare_frame_frequencies}
\alias{compare_frame_frequencies}
\title{Tests all variables in a data frame to check values aren't used too often.}
\usage{
compare_frame_frequencies(data, limit = 0.8)

compare_frame_frequencies(data, limit = 0.8)
}
\arguments{
\item{data}{Data frame of variables to check each column of}

\item{limit}{Max percentage of a variable that can be the same value}
}
\value{
Boolean-vector

Boolean-vector
}
\description{
Tests all variables in a data frame to check values aren't used too often.

Tests all variables in a data frame to check values aren't used too often.
}
