% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscoding_explorative.R
\name{compare_frequencies}
\alias{compare_frequencies}
\title{Vector function - check that no single value of a 
variable is used more than a set limit.}
\usage{
compare_frequencies(values, limit = 0.8)
}
\description{
@param values Vector of values to check
 @param limit The upper limit on how often one value should occur.  Real value [0, 1]. 
 
 @return Boolean 
 @export
}
