% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R, R/miscoding_prescriptive.R
\name{compare_functions}
\alias{compare_functions}
\title{Determines which function to use for validation}
\usage{
compare_functions(value)

compare_functions(value)
}
\arguments{
\item{value}{}
}
\value{
validation function
}
\description{
Determines which function to use for validation

Determines which function to use for validation
}
