% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_vector.R
\name{data.quality.app}
\alias{data.quality.app}
\title{R shiny app for visualizing data quality patterns}
\usage{
data.quality.app(data, group_var)
}
\arguments{
\item{data}{Data frame of observations to check for DQ.}

\item{group_var}{Variable to visualized data quality as a function of.}
}
\description{
Visualizes data quality metrics for  a data set as a function of 
a grouping variable.  The grouping variable can be set as required though 
using a temporal or geographical factor is advised.
}
