% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{data.quality}
\alias{data.quality}
\title{Data quality measures}
\usage{
data.quality(data)
}
\arguments{
\item{data}{Data frame to perform data quality checks on.}
}
\value{
Data frame summarizing data quality metrics.
}
\description{
Perform missing and misscoding checks on supplied data set
}
