% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutual_info.R
\name{discrete.data}
\alias{discrete.data}
\title{Discretize data after encoding factors}
\usage{
discrete.data(data, ...)
}
\arguments{
\item{data}{The data set to be discretized.}

\item{...}{Extra arguments to pass to the infotheo::discretize call.}
}
\description{
Converts non-numeric data to factors and discretizes all features.
}
