% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutual_info.R
\name{encode_values}
\alias{encode_values}
\title{Perform a label encoding on a categorical varaible.}
\usage{
encode_values(values)
}
\arguments{
\item{values}{Vector of values to be encoded.}
}
\value{
Vector of integer values
}
\description{
Perform a label encoding on a categorical varaible.
}
