% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_info_shiny.R
\name{grouped.info.app}
\alias{grouped.info.app}
\title{Interactive app for tracking mutual info.}
\usage{
grouped.info.app(data, group_var)
}
\arguments{
\item{data}{The data set to calculate mutual info from}

\item{group_var}{The variable to group by and use as the x-axis in plots}
}
\value{
Starts a shiny App where variable pairs can be selected and the 
change in mutual information shown.
}
\description{
Interactive app for tracking mutual info.
}
