% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_info_shiny.R
\name{grouped.info.function}
\alias{grouped.info.function}
\title{Internal wrapper to grouped.info.plot}
\usage{
grouped.info.function(data, group_var, Col1, Col2, scale = "Linear")
}
\arguments{
\item{data}{Data frame to be analyzed}

\item{group_var}{The variable to group data by.  Acts as the x-axis in the plot.}

\item{Col1}{A column of data to use for mutual information.}

\item{Col2}{A column of data to use for mutual information}

\item{scale}{Scale type for y-axis - choice of 'Linear' or 'Log'}
}
\value{
A ggplot object.
}
\description{
Internal wrapper to grouped.info.plot
}
