% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{is.missing}
\alias{is.missing}
\title{Vector function - identify which values are 'missing'}
\usage{
is.missing(values, missing_types = MISSING_TYPES)
}
\arguments{
\item{values}{Vector of values to check for missingness}

\item{missing_types}{Vector of values classified as missing.}
}
\value{
Vector of which values are 'missing'
}
\description{
Vector function - identify which values are 'missing'
}
