% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{loc_date_DQ_report}
\alias{loc_date_DQ_report}
\title{Automatically generate a RMD report of data quality issues.}
\usage{
loc_date_DQ_report(data, location.col, date.col, output_path)
}
\arguments{
\item{data}{The data set to be reported on}

\item{location.col}{The column of `data` to treat as a location}

\item{date.col}{The column of `data` to treat as a date,}

\item{output_path}{The location on file path at which to save the report.
NB: '_Loc_Date_Report.html' will be appended}
}
\value{

}
\description{
Automatically generate a RMD report of data quality issues.
}
