% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{missing_DQ_report}
\alias{missing_DQ_report}
\title{Automatically generate a RMD report of data quality issues.}
\usage{
missing_DQ_report(data, output_path)
}
\arguments{
\item{data}{The data set to be reported on}

\item{output_path}{The location on file path at which to save the report.
NB: '_Misising_DQ_Report.html' will be appended}
}
\value{

}
\description{
Automatically generate a RMD report of data quality issues.
}
