% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_mining.R
\name{missing_mine}
\alias{missing_mine}
\title{Analyze each column of a data set that has missing values for patterns of missingness.}
\usage{
missing_mine(data, method = "rpart")
}
\arguments{
\item{data}{Data frame to be analyzed}

\item{method}{`caret::train` method to be applied.}
}
\value{
Vector of ROC values.  Values greater than 0.5 imply some predictve power.
}
\description{
Analyze each column of a data set that has missing values for patterns of missingness.
}
