% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_mining.R
\name{model_roc}
\alias{model_roc}
\title{Calculates model quality via ROC AUC.}
\usage{
model_roc(data, model)
}
\arguments{
\item{data}{Data frame of values to be analyzed.}

\item{model}{`caret` model that has been pre-trained.}
}
\value{
ROC value
}
\description{
Calculates model quality via ROC AUC.
}
