% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutual_info_app.R
\name{mutual.info.app}
\alias{mutual.info.app}
\title{Interactive app for tracking mutual info.}
\usage{
mutual.info.app(data)
}
\arguments{
\item{data}{The data set to calculate mutual info from}
}
\description{
Interactive app for tracking mutual info.
}
