% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutual_info.R
\name{mutual_info}
\alias{mutual_info}
\title{Calculate mutual information between pairs of variables in a data frame}
\usage{
mutual_info(data)
}
\arguments{
\item{data}{Data frame (n by k) of observations}
}
\value{
matrix (k by k) of mutual information.  Diagonal represents total information a varaible has.
}
\description{
Calculate mutual information between pairs of variables in a data frame
}
