% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare2.R, R/miscoding_explorative.R
\name{sqrt_comparison}
\alias{sqrt_comparison}
\title{Function for comparing two values.}
\usage{
sqrt_comparison(l_unique, l_total)

sqrt_comparison(l_unique, l_total)
}
\arguments{
\item{l_unique}{number of unique values}

\item{l_total}{number of observations}
}
\value{
Boolean - FALSE: more unique values than the square root of the number of observations.

Boolean - TRUE: more unique values than the square root of the number of observations.
}
\description{
Function for comparing two values.

Function for comparing two values.
}
